% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPQtree.R
\name{convertPQtree}
\alias{convertPQtree}
\title{Convert linearly encoded \emph{PQ-trees} into a two-dimensional
\emph{PQ-structure}}
\usage{
convertPQtree(rawtree)
}
\arguments{
\item{rawtree}{Data frame with a single column that encodes the structure of
\emph{PQ-trees} as strings. See Details.}
}
\value{
A data frame encoding \emph{PQ-trees} of an ancestral genome
  reconstruction as a two-dimensional \emph{PQ-structure}.

  See the description of the \code{"compgenome"} class in the Details section
  of the \code{\link{checkInfile}} function for more information on the
  output format.
}
\description{
Convert linearly encoded \emph{PQ-trees} of an ancestral genome
reconstruction, for example as output by the software ANGES (Jones \emph{et
al.} 2012), into a two-dimensional \emph{PQ-structure} that can be used as
\code{compgenome} input for the functions \code{\link{computeRearrs}},
\code{\link{summarizeBlocks}}, and \code{\link{genomeRearrPlot}}
}
\details{
\code{rawtree} is a data frame with a single column encoding an ancestral
  genome reconstruction as strings, for example as generated by the software
  ANGES (Chauve & Tannier 2008; Jones \emph{et al.} 2012). The first row
  gives the name of the ancestor (preceded by \code{>}), followed by two rows
  for each ancestral genome segment (CAR). The first row of each set gives
  the CAR ID (preceded by \code{#CAR}), and the second row provides the
  \emph{PQ-tree} structure of that CAR. The children of each node in the
  \emph{PQ-tree} are enclosed by \code{_P} and \code{P_} markups for
  \emph{P-nodes}, or by \code{_Q} and \code{Q_} markups for \emph{Q-nodes}.
  Markers (i.e., ortholog IDs) that belong to a particular node are located
  between its corresponding markups. Markers with reversed orientation are
  preceded by a \code{-} sign. The opening (\code{_P} or \code{_Q}) and
  closing (\code{P_} or \code{Q_}) markups can be nested to allow the
  representation of the hierarchical structure of the \emph{PQ-tree}. For
  details on \emph{PQ-trees} see Booth & Lueker 1976, Chauve & Tannier 2008,
  or the package vignette.

  Example of a \code{rawtree} object, corresponding to
  \code{\link{TOY24_compgenome}}:\cr

  \code{>TOY24}\cr
  \code{#CAR1}\cr
  \code{_Q 1 2 3 4 5 -6 7 8 Q_}\cr
  \code{#CAR2}\cr
  \code{_Q 9 _Q 10 Q_ _Q 11 12 13 Q_ Q_}\cr
  \code{#CAR3}\cr
  \code{_Q 14 Q_}\cr
  \code{#CAR4}\cr
  \code{_Q _P 15 16 17 18 P_ _Q _Q -19 20 21 Q_ _Q -22 -23 24 Q_ Q_ Q_}

  \strong{Important:} The function was designed to work with the ANGES
  \code{*PQTREE*} output file. It was not desiged to work with the
  ancillary \code{*PQRTREE*} or the \code{*DOUBLED*} output files.
}
\section{References}{


  Booth, K.S. & Lueker, G.S. (1976). Testing for the consecutive ones
  property, interval graphs, and graph planarity using \emph{PQ}-Tree
  algorithms. \emph{Journal of Computer and System Sciences}, \strong{13},
  335--379. doi:
  \href{https://doi.org/10.1016/S0022-0000(76)80045-1}{10.1016/S0022-0000(76)80045-1}.

  Chauve, C. & Tannier, E. (2008). A methodological framework for the
  reconstruction of contiguous regions of ancestral genomes and its
  application to mammalian genomes. \emph{PLOS Computational Biology},
  \strong{4}, e1000234. doi:
  \href{https://doi.org/10.1371/journal.pcbi.1000234}{10.1371/journal.pcbi.1000234}.

  Jones, B. R. \emph{et al.} (2012). ANGES: reconstructing ANcestral GEnomeS
  maps. \emph{Bioinformatics}, \strong{28}, 2388--2390. doi:
  \href{https://doi.org/10.1093/bioinformatics/bts457}{10.1093/bioinformatics/bts457}
}

\examples{
\dontrun{

## Read a text file with an ancestral genome reconstruction
## from the software ANGES into R:
MSSYE_PQTREE_HEUR <- read.table("PATH/TO/MSSYE_PQTREE_HEUR",
                                sep = ",", comment.char = "",
                                as.is = TRUE)

## Convert linearly encoded PQ-trees into PQ-structure:
MSSYE_compgenome <- convertPQtree(MSSYE_PQTREE_HEUR)
}

}
\seealso{
\code{\link{checkInfile}}, \code{\link{computeRearrs}},
  \code{\link{summarizeBlocks}}, \code{\link{genomeRearrPlot}}.
}
