% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeRearrPlot.R
\name{genomeRearrPlot}
\alias{genomeRearrPlot}
\title{Genome Rearrangements Plot}
\usage{
genomeRearrPlot(BLOCKS, compgenome, ordfocal, remstr = "", main = "",
  remThld = 0.05, mar = NULL, pad = 0, y0pad = 5,
  uniqueCarColor = TRUE, sortColsBySize = TRUE, plotelem = c(1, 1, 1,
  1, 1), simplifyTags = TRUE, blockwidth = 1, yaxlab = NULL,
  carColors = NULL, carTextColors = NULL, cex.main = 2,
  cex.axis = 1.2, cex.text = 1, font.main = 3, makepdf = FALSE,
  newdev = TRUE, filename = "rearr.pdf", colormodel = "srgb")
}
\arguments{
\item{BLOCKS}{A list of lists generated with the
\code{\link{summarizeBlocks}} function. The top-level list elements of
\code{BLOCKS} are focal genome segments, and the lower-level list elements
contain information on synteny blocks and rearrangements for each focal
genome segment.}

\item{compgenome}{Data frame representing the compared genome (e.g., an
ancestral genome reconstruction, or an extant genome), with the first three
columns \code{$marker}, \code{$orientation}, and \code{$car}, followed by
columns alternating node type and node element. Markers need to be ordered
by their node elements. It must be the same data frame that was used to
generate the list \code{BLOCKS} with the \code{\link{summarizeBlocks}}
function.}

\item{ordfocal}{Character vector with the IDs of the focal genome segments
that will be plotted. Have to match (a subset of) names of the top-level
list elements of \code{BLOCKS}.}

\item{remstr}{String that should be removed from the IDs in \code{ordfocal}
to simplify the y-axis labels. Only relevant when \code{yaxlab = NULL}.}

\item{main}{Title of the plot.}

\item{remThld}{A numeric value between \code{0} (inclusive) and \code{0.5}
(exclusive). Controls whether components of rearrangements that are less
parsimonious to have changed position relative to the alternative
components will be plotted. To plot all components, \code{remThld} needs to
be smaller than \code{remWgt} used in the \code{\link{computeRearrs}}
function.}

\item{mar}{A numerical vector of the form \code{c(bottom, left, top, right)}
that specifies the margins on the four sides of the plot. The default
\code{mar = NULL} sets the margins automatically.}

\item{pad}{A numeric value of \code{0} or greater that sets the amount of
space between all plot margins and the actual plot area.}

\item{y0pad}{A numeric value of \code{0} or greater that sets the amount of
additional space between the bottom plot margin and the bottom plot area.
Setting this value too small may result in some rearrangements for the
bottom-most focal genome segment to be outside the bottom plot area (and
thus to be invisible in the plot).}

\item{uniqueCarColor}{Logical. If \code{TRUE}, CARs are uniquely colored
across all focal genome segments. If \code{FALSE}, CARs are colored
separately for each focal genome segment based on the number of markers per
CAR (forces \code{sortColsBySize = TRUE}).}

\item{sortColsBySize}{Logical. If \code{TRUE}, \code{carColors} and
\code{carTextColors} are assigned based on the number of markers per CAR,
so that the first color is allocated to the largest CAR.}

\item{plotelem}{A numerical vector of the form \code{c(nor, bid, bor, eid,
rea)} that determines which synteny block information is visualized.
\code{nor} is the alignment orientation of the node, \code{bid} is the
block ID, \code{bor} is the block orientation within its node, \code{eid}
is the element ID within its node, and \code{rea} are rearrangements. The
information is plotted when the value is \code{1}, and omitted when
\code{0}.}

\item{simplifyTags}{Logical. If \code{TRUE}, duplicated rearrangement tags,
if any, are excluded from the plot. Note that this will work properly only
when \code{remWgt} used in the \code{\link{computeRearrs}} function was set
to a value \code{>0}.}

\item{blockwidth}{A numeric value that specifies the relative width of
synteny blocks (i.e., columns) in the plot.}

\item{yaxlab}{Annotations of the y-axis. Must be the same length as
\code{ordfocal}. The default \code{yaxlab = NULL} uses as annotations the
names in \code{ordfocal}, simplified through removal of the string in
\code{remstr}.}

\item{carColors}{Character vector with the color names used for coloring
CARs. If the number of CARs is greater than \code{length(carColors)},
remaining CARs are colored in grayscale. \code{carColors} must have the
same length as \code{carTextColors}.}

\item{carTextColors}{Character vector with the color names used for
coloring CAR IDs. \code{carTextColors} must have the same length as
\code{carColors}.}

\item{cex.main}{Numerical value that specifies the magnification of the main
title.}

\item{cex.axis}{Numerical value that specifies the magnification of the axis
annotation.}

\item{cex.text}{Numerical value that specifies the magnification of text
within the plot area.}

\item{font.main}{Font for the title of the plot. \code{1} corresponds to
plain text, \code{2} to bold face, \code{3} to italic, and \code{4} to bold
italic.}

\item{makepdf}{Logical. Save plot as PDF. See \code{filename} and
\code{colormodel}.}

\item{newdev}{Logical. Opens a new default graphics device (but not
"RStudioGD") via \code{dev.new}. Only relevant when \code{makepdf = FALSE}.}

\item{filename}{Character string that gives the name of the PDF file when
\code{makepdf = TRUE}.}

\item{colormodel}{Character string that gives the color model for the PDF
when \code{makepdf = TRUE}. Allowed values are \code{"srgb"},
\code{"cmyk"}, \code{"gray"}, or \code{"grey"}.}
}
\value{
A plot to the default graphics device (but not "RStudioGD") or a PDF
  file.

  The plot visualizes the data contained in \code{BLOCKS} for each focal
  genome segment in \code{ordfocal}, arranged along the y-axis. Each synteny
  block is represented by a column (corresponding to rows in \code{BLOCKS}),
  and information on each block is visualized in rows (corresponding to
  columns in \code{BLOCKS}). Note that separate blocks are also generated
  when the hierarchical structure of the underlying \emph{PQ-tree} changes,
  therefore not all column boundaries are caused by a rearrangement. Rows
  provide information on the structure of each \emph{PQ-tree} and its
  alignment to the focal genome, and whether blocks are part of different
  classes of rearrangements. For details on \emph{PQ-trees} see the
  description of the \code{"compgenome"} class in the Details section of the
  \code{\link{checkInfile}} function, Booth & Lueker 1976, Chauve & Tannier
  2008, or the package vignette.

  For each focal genome segment, the top row gives the number of markers
  within each block, followed by a row that gives the IDs of the CARs. The
  remaining rows are optional and are controlled by the values in
  \code{plotelem}.

  Up to four rows (depending on the values in the argument \code{plotelem})
  describe the structure of each \emph{PQ-tree} and its alignment to the
  focal genome, and are stacked for each level of the \emph{PQ-tree}
  hierarchy. The first row (\code{nor}) gives the alignment orientation of
  the \emph{PQ-tree} node to the focal genome, with white rectangles and
  \code{"+"} indicating ascending (i.e., standard), and black rectangles and
  \code{"-"} indicating descending (i.e., inverted) alignment. Nodes that
  have no alignment direction (e.g., single-marker nodes) are in gray, and
  \emph{P-nodes} (which have no fixed ordering and thus no direction) have
  gray shaded rectangles. The second row (\code{bid}) gives the block ID. For
  \emph{Q-nodes}, IDs are consecutive and start at \code{1}, separately for
  each node and each hierarchy level, and reflect the order of synteny
  blocks. Identical IDs mean that blocks might be joined, but are split
  either by an insertion of another CAR or because of a change in the
  hierarchical structure of the underlying \emph{PQ-tree}. Block IDs with
  \code{".1"} or \code{".2"} suffixes indicate rearrangements that may arise
  by either an inversion or a translocation between adjacent blocks, but that
  were classified as translocation for the sake of parsimony. For
  \emph{P-nodes}, the \code{bid} row is empty unless the node is part of a
  rearrangement, in which case IDs indicate different rearrangements, but not
  block order. The third row (\code{bor}) gives the block orientation within
  its node. It has the same color and symbol coding as \code{nor} above. For
  example, a \code{"-"} block within a \code{"+"} node indicates either an
  inversion or a translocation between adjacent blocks. The fourth row
  (\code{eid}) gives the range of element IDs for each block within its node
  and for its level of hierarchy. These IDs correspond to the node elements
  in the odd columns of \code{compgenome} (note that some IDs within blocks
  or in-between might be missing when markers in the compared genome are
  absent from the focal genome).

  The final set of rows (\code{rea}) indicates whether blocks are part of
  different classes of rearrangements. Horizontal lines that are at identical
  height denote the same rearrangement (potentially disrupted by inserted
  CARs). Green are translocations between CARs between focal segments (TLBS);
  blue are translocations between CARs within focal segments (TLWS); purple
  are translocations within CARs within focal segments (TLWC); maroon are
  inversions within CARs within focal segments (IV). Inversions that involve
  only a single marker (i.e., markers with switched orientation) are
  indicated by a short vertical rather than a horizontal line. Lighter
  coloration denotes smaller weights for rearrangement tags in the respective
  matrices in \code{BLOCKS}. Unless the argument \code{remThld} is set to a
  value smaller than that of \code{remWgt} used in the
  \code{\link{computeRearrs}} function, only lines for blocks that are more
  parsimonious to have changed position relative to alternative blocks are
  plotted. If \code{simplifyTags = FALSE}, all tags for TLWS and TLWC will be
  plotted for completeness, i.e., including those that are duplicated due to
  the functioning of the underlying algorithm in \code{\link{computeRearrs}}.
}
\description{
Generate a plot that shows synteny blocks between a focal and a compared
genome in columns, and information on their alignment and rearrangements in
rows, for a given set of focal genome segments
}
\details{
Parameters \code{BLOCKS}, \code{compgenome}, and \code{ordfocal} need to be
specified, all other parameters have default or automatic settings.

When \code{makepdf = TRUE} or \code{newdev = TRUE}, the width and height of
the graphic will be set automatically. The dimensions are determined in
inches, thus \code{makepdf = FALSE} and \code{newdev = TRUE} will produce an
error or not work correctly when the default units of the default graphics
device are not inches (such as \code{\link[grDevices]{bmp}},
\code{\link[grDevices]{jpeg}}, \code{\link[grDevices]{png}}, or
\code{\link[grDevices]{tiff}}). This can be avoided by setting the default
graphics device to one that has inches as default units. Setting both
\code{makepdf = FALSE} and \code{newdev = FALSE} will allow to specify
alternative, user-defined dimensions of the graphic. See examples below.

Colors are assigned to CARs by size, unless \code{sortColsBySize = FALSE}.
When \code{carColors = NULL}, 47 easily distinguishable default colors are
used for coloring CARs. The first 14 colors are color blindness friendly and
were obtained from
\href{http://mkweb.bcgsc.ca/biovis2012}{mkweb.bcgsc.ca/biovis2012}.
\code{carTextColors} are either black or white dependent on the hue of the
default \code{carColors}.
}
\section{References}{


  Booth, K.S. & Lueker, G.S. (1976). Testing for the consecutive ones
  property, interval graphs, and graph planarity using \emph{PQ}-Tree
  algorithms. \emph{Journal of Computer and System Sciences}, \strong{13},
  335--379. doi:
  \href{https://doi.org/10.1016/S0022-0000(76)80045-1}{10.1016/S0022-0000(76)80045-1}.

  Chauve, C. & Tannier, E. (2008). A methodological framework for the
  reconstruction of contiguous regions of ancestral genomes and its
  application to mammalian genomes. \emph{PLOS Computational Biology},
  \strong{4}, e1000234. doi:
  \href{https://doi.org/10.1371/journal.pcbi.1000234}{10.1371/journal.pcbi.1000234}.
}

\examples{
SYNT<-computeRearrs(TOY24_focalgenome, TOY24_compgenome, doubled = TRUE)
BLOCKS<-summarizeBlocks(SYNT, TOY24_focalgenome, TOY24_compgenome,
                        c("1","2","3"))

genomeRearrPlot(BLOCKS, TOY24_compgenome, c("1", "2", "3"), main = "TOY24")

\dontrun{

## make PDF (automatically determine the width and height of the graphic)
genomeRearrPlot(BLOCKS, TOY24_compgenome, c("1", "2", "3"), main = "TOY24",
                makepdf = TRUE, newdev = FALSE, filename = "rearr.pdf")

## make PDF (default dimensions, i.e., square format)
pdf(file = "rearr.pdf")
genomeRearrPlot(BLOCKS, TOY24_compgenome, c("1", "2", "3"), main = "TOY24",
                makepdf = FALSE, newdev = FALSE)
dev.off()

## plot in R Studio window
op <- options(device = "RStudioGD")
genomeRearrPlot(BLOCKS, TOY24_compgenome, c("1", "2", "3"), main = "TOY24",
                newdev = FALSE)
options(op)

## make EPS, and set user-specified dimensions
setEPS()
postscript("rearr.eps", width=4.5,height=6.0,pointsize=9)
genomeRearrPlot(BLOCKS, TOY24_compgenome, c("1", "2"), main = "TOY24",
                pad = 1, y0pad = 1, makepdf = FALSE, newdev = FALSE)
dev.off()
}

}
\seealso{
\code{\link{checkInfile}}, \code{\link{computeRearrs}},
  \code{\link{summarizeBlocks}}, \code{\link{genomeImagePlot}}. For more
  information about arguments that are passed to other functions, see
  \code{\link[grDevices]{dev.new}}, \code{\link[grDevices]{pdf}},
  \code{\link[graphics]{plot}}, \code{\link[graphics]{par}}.
}
