% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drosophila_data.R
\docType{data}
\name{MEL_markers}
\alias{MEL_markers}
\title{Focal genome map (i.e., \code{focalgenome}) of \emph{Drosophila melanogaster}}
\format{A data frame with markers in rows and six columns that specify the
  structure of the \emph{D. melanogaster} focal genome map:
\tabular{ll}{
  \code{$marker} \tab unique ortholog ID of marker \cr
  \code{$scaff} \tab genome segment where marker is located \cr
  \code{$start} \tab start position of marker on genome segment \cr
  \code{$end} \tab end position of marker on genome segment \cr
  \code{$strand} \tab reading direction of marker \cr
  \code{$name} \tab marker name \cr
}}
\usage{
MEL_markers
}
\description{
A data set representing the \emph{D. melanogaster} genome with 13,818 markers
on 12 genome segments (i.e., chromosomes and scaffolds). The genome map was
generated based on publicly available genome assemblies downloaded from
Ensemble Release 91
\href{http://dec2017.archive.ensembl.org}{(http://dec2017.archive.ensembl.org)}
and Ensemble Metazoa Release 37
\href{http://oct2017-metazoa.ensembl.org}{(http://oct2017-metazoa.ensembl.org)}.
Orthologs were identified with OMA standalone v2.2.0 (Altenhoff \emph{et al.}
2015). See the package vignette for details.
}
\section{References}{


  Altenhoff, A.M. \emph{et al.} (2015). The OMA orthology database in 2015:
  function predictions, better plant support, synteny view and other
  improvements. \emph{Nucleic Acids Research}, \strong{43}, D240--D249. doi:
  \href{https://doi.org/10.1093/nar/gku1158}{10.1093/nar/gku1158}.
}

\seealso{
\code{\link{computeRearrs}}, \code{\link{SIM_markers}},
  \code{\link{YAK_markers}}, \code{\link{MSSYE_PQTREE_HEUR}}
}
\keyword{datasets}
