% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeRearrs.R
\name{summarizeRearrs}
\alias{summarizeRearrs}
\title{Summarize Rearrangements}
\usage{
summarizeRearrs(SYNT, focalgenome, compgenome, ordfocal, remWgt = 0.05,
  remThld = 0)
}
\arguments{
\item{SYNT}{A list of matrices that store data on different classes of
rearrangements and additional information. \code{SYNT} must have been
generated with the \code{\link{computeRearrs}} function (optionally
filtered with the \code{\link{filterRearrs}} function).}

\item{focalgenome}{Data frame representing the focal genome, containing the
mandatory columns \code{$marker}, \code{$scaff}, \code{$start},
\code{$end}, and \code{$strand}, and optional further columns. Markers need
to be ordered by their map position.}

\item{compgenome}{Data frame representing the compared genome (e.g., an
ancestral genome reconstruction, or an extant genome), with the first three
columns \code{$marker}, \code{$orientation}, and \code{$car}, followed by
columns alternating node type and node element. Markers need to be ordered
by their node elements. \code{compgenome} must be the same data frame that
was used to generate the list \code{SYNT} with the
\code{\link{computeRearrs}} function.}

\item{ordfocal}{Character vector with the IDs of the focal genome segments
that will be summarized. Have to match (a subset of) IDs in
\code{focalgenome$scaff}.}

\item{remWgt}{A numeric value between \code{0} (inclusive) and \code{0.5}
(exclusive). Needs to match the value for \code{remWgt} used in the
\code{\link{computeRearrs}} function.}

\item{remThld}{A numeric value between \code{0} (inclusive) and \code{0.5}
(exclusive). Controls whether breakpoints for components of rearrangements
that are less parsimonious to have changed position relative to the
alternative components will be output. To output all breakpoints,
\code{remThld} needs to be smaller than \code{remWgt} used in the
\code{\link{computeRearrs}} function.}
}
\value{
A matrix with the number of identified nonsyntenic moves, syntenic
  moves, inversions, and breakpoints in columns, for the set of focal genome
  segments in \code{ordfocal} in rows.
}
\description{
For each focal genome segment, summarize the number and type of rearrangement
events and the number of breakpoints
}
\details{
Only rearrangements that have components tagged with values larger than
  \code{remWgt} will be counted. For proper functioning, \code{remWgt} should
  correspond to the value that has been used to generate \code{SYNT}. The
  number of nonsyntenic moves is computed as the maximum of class I and class
  II nonsyntenic moves per focal genome segment.

  The number of breakpoints is computed based on the
  \code{\link{getBreakpoints}} function. To include the breakpoint of origin
  for nonsyntenic and syntenic moves in the estimate, \code{remThld} needs to
  be set to zero (which is the default). Note that this may nevertheless
  underestimate the number of breakpoints as the location of origin is not
  determined for all rearrangements. When the input is a filtered version of
  \code{SYNT} (i.e., filtered with the \code{\link{filterRearrs}} function),
  the number of breakpoints may be overestimated. This can be prevented by
  increasing the value of \code{remThld} to match the value of \code{remWgt}.
  However, this may then underestimate the number of breakpoints as some
  breakpoints of origin will not be counted. Breakpoints that fall on
  identical positions are only counted once.
}
\examples{
SYNT <- computeRearrs(TOY24_focalgenome, TOY24_compgenome, doubled = TRUE)

summarizeRearrs(SYNT, TOY24_focalgenome, TOY24_compgenome, c("1","2","3"))

}
\seealso{
\code{\link{computeRearrs}}, \code{\link{filterRearrs}},
  \code{\link{getBreakpoints}}.
}
