% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOY24_data.R
\docType{data}
\name{TOY24_focalgenome}
\alias{TOY24_focalgenome}
\title{\code{focalgenome} example for 24 markers, created for illustrative purposes}
\format{A data frame with markers in rows and five columns that specify the
  structure of a focal genome map:
\tabular{ll}{
  \code{$marker} \tab unique ortholog ID of marker \cr
  \code{$scaff} \tab genome segment where marker is located \cr
  \code{$start} \tab start position of marker on genome segment \cr
  \code{$end} \tab end position of marker on genome segment \cr
  \code{$strand} \tab reading direction of marker \cr
}}
\usage{
TOY24_focalgenome
}
\description{
A data set illustrating a focal (extant) genome map with 24 markers on three
genome segments (e.g., scaffolds).
}
\seealso{
\code{\link{computeRearrs}}, \code{\link{TOY24_compgenome}}
}
\keyword{datasets}
